WIP
Fixed ints with void pointers, now working on implementing info nested structure. Running into issues with head pointer being reset due to nesting dictionaries. Am going to need to figure out a new way of handling the head pointer so that the head isn't lost. Some kind of memory for and logic surrounding head pointer when dictionary is called? We have to handle at least triple nested dictionaries so this should be a fun challenge.